Given by Codecademy
This is an in-browser game that the user plays against a CPU.
The secret number is generated by Math.random and is between 0 and 9.
The computer guesses a random number between 0 and 9, while the user can input a number between 0 and 9.
Whoever guesses closest to the secret number is the winner and gets 1 point added to their score. The user can click next round to continue the game.